+++ extensions/gitlab/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `Merge Request Menu` command
 - Add `Issues Menu` command
 - Add group filter dropdown for epics
+- `My Groups` only show one level at a time
 
 ## [Search Scope] - 2023-04-27